Initial commit of prebuilt w64 cross-toolchain.
This patch adds the binaries of a new prebuilt cross-toolchain
that can be used to build Win32 and Win64 binaries on Linux.

This is to be used to build official Windows binaries for the
Android SDK, in particular to get a working emulator + GPU emulation
libraries.

See README for details.

BUG=15402623

Change-Id: I090543141b33babf99b9ecaea3dbdd5c3cc03d8e